refactor: merge footer and share button - Combine detail footer and share button into ContentFooter - Simplify post/project footer usage to URL-only sharing
